+++ src/plug/schedule.js
@@ -13,7 +13,8 @@ class PlugSchedule extends Schedule {
    *
    * Sends `schedule.add_rule` command and returns rule id.
    * @param  {Object}        options
-   * @param  {boolean}       options.powerState
+   * @param  {boolean}      [options.powerState]
+   * @param  {Object}       [options.dimmer] dimmer data (dimmable plugs only)
    * @param  {(Date|number)} options.start  Date or number of minutes
    * @param  {number[]}     [options.daysOfWeek]  [0,6] = weekend, [1,2,3,4,5] = weekdays
    * @param  {string}       [options.name]
@@ -21,7 +22,7 @@ class PlugSchedule extends Schedule {
    * @param  {SendOptions}  [sendOptions]
    * @return {Promise<Object, ResponseError>} parsed JSON response
    */
-  async addRule ({ powerState, start, daysOfWeek, name = '', enable = true }, sendOptions) {
+  async addRule ({ powerState, dimmer, start, daysOfWeek, name = '', enable = true }, sendOptions) {
     const rule = Object.assign({
       sact: (powerState ? 1 : 0),
       name,
@@ -29,6 +30,11 @@ class PlugSchedule extends Schedule {
       emin: 0,
       etime_opt: -1
     }, createScheduleRule({ start, daysOfWeek }));
+
+    if (dimmer !== undefined) {
+      rule.sact = 3;
+      rule.s_dimmer = dimmer;
+    }
 
     return Schedule.prototype.addRule.call(this, rule, sendOptions); // super.addRule(rule); // workaround babel bug
   }
@@ -38,7 +44,8 @@ class PlugSchedule extends Schedule {
    * Sends `schedule.edit_rule` command and returns rule id.
    * @param  {Object}        options
    * @param  {string}        options.id
-   * @param  {boolean}       options.powerState
+   * @param  {boolean}      [options.powerState]
+   * @param  {Object}       [options.dimmer] dimmer data (dimmable plugs only)
    * @param  {(Date|number)} options.start  Date or number of minutes
    * @param  {number[]}     [options.daysOfWeek]  [0,6] = weekend, [1,2,3,4,5] = weekdays
    * @param  {string}       [options.name]    [description]
@@ -46,7 +53,7 @@ class PlugSchedule extends Schedule {
    * @param  {SendOptions}  [sendOptions]
    * @return {Promise<Object, ResponseError>} parsed JSON response
    */
-  async editRule ({ id, powerState, start, daysOfWeek, name = '', enable = true }, sendOptions) {
+  async editRule ({ id, powerState, dimmer, start, daysOfWeek, name = '', enable = true }, sendOptions) {
     const rule = Object.assign({
       id,
       sact: (powerState ? 1 : 0),
@@ -55,6 +62,11 @@ class PlugSchedule extends Schedule {
       emin: 0,
       etime_opt: -1
     }, createScheduleRule({ start, daysOfWeek }));
+
+    if (dimmer !== undefined) {
+      rule.sact = 3;
+      rule.s_dimmer = dimmer;
+    }
 
     return Schedule.prototype.editRule.call(this, rule, sendOptions); // super.editRule(rule); // workaround babel bug
   }